+++ src/main/java/com/example/appointback/controller/MedServiceRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MedServiceRepository extends CrudRepository<MedicalService, Long> {
+public interface MedServiceRepository extends CrudRepository<MedicalService, Short> {
 
     List<MedicalService> findAll();
 }
